- be-develop 브랜치 develop 브랜치로 병합

+++ packages/server/src/board/board.service.ts
@@ -8,25 +8,50 @@ import { UpdateBoardDto } from './dto/update-board.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Board } from './entities/board.entity';
 import { Repository } from 'typeorm';
+import { Image } from './entities/image.entity';
+import { encryptAes } from 'src/utils/aes.util';
+import { User } from 'src/auth/entities/user.entity';
+import { UserDataDto } from './dto/user-data.dto';
+import * as AWS from 'aws-sdk';
+import { awsConfig, bucketName } from 'src/config/aws.config';
+import { v1 as uuid } from 'uuid';
 
 @Injectable()
 export class BoardService {
 	constructor(
 		@InjectRepository(Board)
-		private boardRepository: Repository<Board>,
+		private readonly boardRepository: Repository<Board>,
+		@InjectRepository(Image)
+		private readonly imageRepository: Repository<Image>,
+		@InjectRepository(User)
+		private readonly userRepository: Repository<User>,
 	) {}
 
-	async create(createBoardDto: CreateBoardDto): Promise<Board> {
-		const { title, content, author } = createBoardDto;
+	async createBoard(
+		createBoardDto: CreateBoardDto,
+		userData: UserDataDto,
+		files: Express.Multer.File[],
+	): Promise<Board> {
+		const { title, content } = createBoardDto;
+
+		const user = await this.userRepository.findOneBy({ id: userData.userId });
+
+		const images: Image[] = [];
+		for (const file of files) {
+			const image = await this.uploadFile(file);
+			images.push(image);
+		}
 
 		const board = this.boardRepository.create({
 			title,
-			content,
-			author,
+			content: encryptAes(content), // AES 암호화하여 저장
+			user,
+			images,
 		});
-		const created: Board = await this.boardRepository.save(board);
+		const createdBoard: Board = await this.boardRepository.save(board);
 
-		return created;
+		createdBoard.user.password = undefined; // password 제거하여 반환
+		return createdBoard;
 	}
 
 	async findAllBoards(): Promise<Board[]> {
@@ -35,7 +60,9 @@ export class BoardService {
 	}
 
 	async findAllBoardsByAuthor(author: string): Promise<Board[]> {
-		const boards = await this.boardRepository.findBy({ author });
+		const boards = await this.boardRepository.findBy({
+			user: { nickname: author },
+		});
 		return boards;
 	}
 
@@ -47,8 +74,22 @@ export class BoardService {
 		return found;
 	}
 
-	async updateBoard(id: number, updateBoardDto: UpdateBoardDto) {
+	async updateBoard(
+		id: number,
+		updateBoardDto: UpdateBoardDto,
+		userData: UserDataDto,
+	) {
 		const board: Board = await this.findBoardById(id);
+
+		// 게시글 작성자와 수정 요청자가 다른 경우
+		if (board.user.id !== userData.userId) {
+			throw new BadRequestException('You are not the author of this post');
+		}
+
+		// updateBoardDto.content가 존재하면 AES 암호화하여 저장
+		if (updateBoardDto.content) {
+			updateBoardDto.content = encryptAes(updateBoardDto.content);
+		}
 
 		const updatedBoard: Board = await this.boardRepository.save({
 			...board,
@@ -57,21 +98,89 @@ export class BoardService {
 		return updatedBoard;
 	}
 
-	async patchLike(id: number): Promise<Partial<Board>> {
+	async patchLike(id: number, userData: UserDataDto): Promise<Partial<Board>> {
 		const board = await this.findBoardById(id);
-		board.like_cnt += 1;
-		await this.boardRepository.save(board);
-		return { like_cnt: board.like_cnt };
+		if (board.likes.find((user) => user.id === userData.userId)) {
+			throw new BadRequestException('You already liked this post');
+		}
+
+		const user = await this.userRepository.findOneBy({ id: userData.userId });
+		if (!user) {
+			throw new NotFoundException(`Not found user with id: ${userData.userId}`);
+		}
+
+		board.likes.push(user);
+		board.like_cnt = board.likes.length;
+
+		const updatedBoard = await this.boardRepository.save(board);
+
+		return { like_cnt: updatedBoard.like_cnt };
 	}
 
-	async patchUnlike(id: number): Promise<Partial<Board>> {
+	async patchUnlike(
+		id: number,
+		userData: UserDataDto,
+	): Promise<Partial<Board>> {
 		const board = await this.findBoardById(id);
-		board.like_cnt -= 1;
-		await this.boardRepository.save(board);
-		return { like_cnt: board.like_cnt };
+		if (!board.likes.find((user) => user.id === userData.userId)) {
+			throw new BadRequestException('You have not liked this post');
+		}
+
+		const user = await this.userRepository.findOneBy({ id: userData.userId });
+		if (!user) {
+			throw new NotFoundException(`Not found user with id: ${userData.userId}`);
+		}
+
+		board.likes = board.likes.filter((user) => user.id !== userData.userId);
+		board.like_cnt = board.likes.length;
+
+		const updatedBoard = await this.boardRepository.save(board);
+		return { like_cnt: updatedBoard.like_cnt };
 	}
 
-	async deleteBoard(id: number): Promise<void> {
+	async deleteBoard(id: number, userData: UserDataDto): Promise<void> {
+		const board: Board = await this.findBoardById(id);
+
+		// 게시글 작성자와 삭제 요청자가 다른 경우
+		if (board.user.id !== userData.userId) {
+			throw new BadRequestException('You are not the author of this post');
+		}
+
 		const result = await this.boardRepository.delete({ id });
+	}
+
+	async uploadFile(file: Express.Multer.File): Promise<Image> {
+		if (!file.mimetype.includes('image')) {
+			throw new BadRequestException('Only image files are allowed');
+		}
+
+		const { mimetype, buffer, size } = file;
+
+		const filename = uuid();
+
+		// AWS.config.update(awsConfig);
+		// const result = await new AWS.S3().putObject({
+		// 	Bucket: bucketName,
+		// 	Key: filename,
+		// 	Body: buffer,
+		// 	ACL: 'public-read',
+		// })
+		// .promise();
+		// console.log(result);
+		// const downloaded = await new AWS.S3()
+		// 	.getObject({
+		// 		Bucket: bucketName,
+		// 		Key: filename,
+		// 	})
+		// 	.promise();
+		// console.log(downloaded);
+
+		const updatedImage = await this.imageRepository.save({
+			mimetype,
+			filename,
+			size,
+		});
+
+		return updatedImage;
 	}
 }
